Restructure util functions
